+++ packages/cli/src/commands/index.ts
@@ -133,9 +133,7 @@ if (process.env.FF_GUIDANCE_MODE) {
 
 commandsStructs.push(metricsCommand);
 
-if (process.env.FF_CONNEX_ENABLED) {
-  commandsStructs.push(connexCommand);
-}
+commandsStructs.push(connexCommand);
 
 export function getCommandAliases(command: Pick<Command, 'name' | 'aliases'>) {
   return [command.name].concat(command.aliases);