+++ packages/build-utils/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @vercel/build-utils
 
+## 13.25.0
+
+### Minor Changes
+
+- fb0cb8d: Add normalized entrypoint detector for runtime builders.
+- 4fc110b: [services] add preDeployCommand for experimentalServices
+
+## 13.24.0
+
+### Minor Changes
+
+- d874af6: Add support for env vars injection that reference other services in `services` with an explicit `env` configuration.
+
 ## 13.23.0
 
 ### Minor Changes