+++ packages/gatsby-plugin-vercel-builder/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @vercel/gatsby-plugin-vercel-builder
 
+## 2.2.5
+
+### Patch Changes
+
+- Updated dependencies [fb0cb8d]
+- Updated dependencies [4fc110b]
+  - @vercel/build-utils@13.25.0
+
+## 2.2.4
+
+### Patch Changes
+
+- Updated dependencies [d874af6]
+  - @vercel/build-utils@13.24.0
+
 ## 2.2.3
 
 ### Patch Changes